modifying function for multiple clients
